Refactored  wrt exceptions usage
