+++ beets/ui/__init__.py
@@ -39,6 +39,7 @@ from beets.util import as_string
 from beets.autotag import mb
 from beets.dbcore import query as db_query
 from beets.dbcore import db
+from beets.dbcore import dbcore_exceptions
 import confuse
 
 # On Windows platforms, use colorama to support "ANSI" terminal colors.
@@ -1318,7 +1319,7 @@ def main(args=None):
     except confuse.ConfigError as exc:
         log.error('configuration error: {0}', exc)
         sys.exit(1)
-    except db_query.InvalidQueryError as exc:
+    except dbcore_exceptions.InvalidQueryError as exc:
         log.error('invalid query: {0}', exc)
         sys.exit(1)
     except OSError as exc:
@@ -1330,7 +1331,7 @@ def main(args=None):
     except KeyboardInterrupt:
         # Silently ignore ^C except in verbose mode.
         log.debug('{}', traceback.format_exc())
-    except db.DBAccessError as exc:
+    except dbcore_exceptions.DBAccessError as exc:
         log.error(
             'database access error: {0}\n'
             'the library file might have a permissions problem',
