+++ beetsplug/smartplaylist.py
@@ -22,7 +22,8 @@ from beets.util import (mkdirall, normpath, sanitize_path, syspath,
                         bytestring_path, path_as_posix)
 from beets.library import Item, Album, parse_query_string
 from beets.dbcore import OrQuery
-from beets.dbcore.query import MultipleSort, ParsingError
+from beets.dbcore.query import MultipleSort
+from beets.dbcore.dbcore_exceptions import ParsingError
 import os
 
 try:
