+++ test/test_query.py
@@ -27,8 +27,8 @@ from test import helper
 import beets.library
 from beets import dbcore
 from beets.dbcore import types
-from beets.dbcore.query import (NoneQuery, ParsingError,
-                                InvalidQueryArgumentValueError)
+from beets.dbcore.query import NoneQuery
+from beets.dbcore.dbcore_exceptions import ParsingError, InvalidQueryArgumentValueError
 from beets.library import Library, Item
 from beets import util
